# frontend/openapi/asbl-club-api.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "asbl.club API",
    "description" : "Public read API exposing associations and their public events. Reads are open with CORS for showcase sites. Everything else requires a JWT access token from POST /api/v1/auth/login, sent as a Bearer token.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/api/v1/guest-bookings" : {
      "post" : {
        "tags" : [ "Guest bookings" ],
        "summary" : "Book a seat on a public event, without an account",
        "operationId" : "bookAsGuest",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GuestBookRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Booked; the secret link to pay and to see the ticket is in the body and was emailed",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GuestBooked"
                }
              }
            }
          },
          "404" : {
            "description" : "No such public event, or a ticket of another event",
            "content" : {
              "application/problem+json" : { }
            }
          },
          "409" : {
            "description" : "Sold out (SOLD_OUT), or the association can't receive payments yet (PAYMENTS_DISABLED)",
            "content" : {
              "application/problem+json" : { }
            }
          }
        }
      }
    },
    "/api/v1/guest-bookings/{token}/checkout" : {
      "post" : {
        "tags" : [ "Guest bookings" ],
        "summary" : "Start (or resume) paying for a guest's booking",
        "operationId" : "startGuestCheckout",
        "parameters" : [ {
          "name" : "token",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ready to pay",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Checkout"
                }
              }
            }
          },
          "409" : {
            "description" : "Nothing to pay (NOTHING_TO_PAY), not paid in time (BOOKING_EXPIRED), or the association can't receive payments yet (PAYMENTS_DISABLED)",
            "content" : {
              "application/problem+json" : { }
            }
          },
          "502" : {
            "description" : "The payment provider couldn't be reached",
            "content" : {
              "application/problem+json" : { }
            }
          }
        }
      }
    },
    "/api/v1/guest-bookings/{token}/cancel" : {
      "post" : {
        "tags" : [ "Guest bookings" ],
        "summary" : "Cancel a guest's paid ticket and be refunded",
        "operationId" : "cancelGuestBooking",
        "parameters" : [ {
          "name" : "token",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Cancelled and refunded",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GuestBooking"
                }
              }
            }
          },
          "409" : {
            "description" : "The delay has passed or tickets aren't refundable (CANCELLATION_CLOSED), or not a paid unused ticket of an event still on (NOT_CANCELLABLE)",
            "content" : {
              "application/problem+json" : { }
            }
          },
          "502" : {
            "description" : "The payment provider couldn't be reached",
            "content" : {
              "application/problem+json" : { }
            }
          }
        }
      }
    },
...
    "/api/v1/guest-bookings/{token}" : {
      "get" : {
        "tags" : [ "Guest bookings" ],
        "summary" : "A guest's booking, through its secret link",
        "operationId" : "getGuestBooking",
        "parameters" : [ {
          "name" : "token",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Unknown link",
            "content" : {
              "application/problem+json" : { }
            }
          }
        }
      }
    },
    "/api/v1/events/{eventId}" : {
      "get" : {
        "tags" : [ "Public" ],
        "summary" : "Get a public event with its association, location and tickets",
        "operationId" : "getEvent",
        "parameters" : [ {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicEvent"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/asbls/{slug}" : {
      "get" : {
        "tags" : [ "Public" ],
        "summary" : "Get an association by slug",
        "operationId" : "getAsbl",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AsblResource"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Checkout" : {
        "required" : [ "amount", "clientSecret", "currency", "publishableKey", "stripeAccount" ],
        "type" : "object",
        "properties" : {
          "publishableKey" : {
            "type" : "string"
          },
          "stripeAccount" : {
            "type" : "string",
            "description" : "The association's Stripe account (payments go straight to it)"
          },
          "clientSecret" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "number",
            "description" : "Amount in euros"
          },
          "currency" : {
            "type" : "string"
          }
        }
      },
...
      "GuestBookRequest" : {
        "required" : [ "email", "eventId", "language", "name", "ticketCategoryId" ],
        "type" : "object",
        "properties" : {
          "eventId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ticketCategoryId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "email" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string",
            "format" : "email"
          },
          "language" : {
            "minLength" : 1,
            "pattern" : "fr|nl|en",
            "type" : "string"
          }
        }
      },
      "GuestBooked" : {
        "required" : [ "accessToken", "booking" ],
        "type" : "object",
        "properties" : {
          "accessToken" : {
            "type" : "string",
            "description" : "Opens the booking: /tickets/{accessToken}. Shown once."
          },
          "booking" : {
            "$ref" : "#/components/schemas/GuestBooking"
          }
        }
      },
      "GuestBooking" : {
        "required" : [ "amount", "asblName", "currency", "email", "eventId", "eventTitle", "name", "startsAt", "status", "ticketLabel" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "enum" : [ "RESERVED", "PAID", "CONFIRMED", "ATTENDED", "CANCELLED", "REFUNDED", "EXPIRED" ]
          },
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "number",
            "description" : "Amount in euros"
          },
          "currency" : {
            "type" : "string"
          },
          "eventId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "eventTitle" : {
            "type" : "string"
          },
          "startsAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "location" : {
            "type" : "string"
          },
          "asblName" : {
            "type" : "string"
          },
          "ticketLabel" : {
            "type" : "string"
          },
          "ticketCode" : {
            "type" : "string",
            "description" : "The code in the ticket's QR code; present once the booking is paid"
          },
          "checkedInAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "cancellableUntil" : {
            "type" : "string",
            "description" : "Until when this ticket can still be cancelled and refunded; absent: it can't",
            "format" : "date-time"
          }
        }
      },
...
      "AsblResource" : {
        "required" : [ "denomination", "slug" ],
        "type" : "object",
        "properties" : {
          "slug" : {
            "type" : "string"
          },
          "denomination" : {
            "type" : "string"
          }
        }
      },
...
      "PublicEvent" : {
        "required" : [ "asbl", "id", "startsAt", "tickets", "title" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "startsAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "location" : {
            "type" : "string"
          },
          "asbl" : {
            "$ref" : "#/components/schemas/AsblResource"
          },
          "cancellableUntil" : {
            "type" : "string",
            "description" : "Until when buyers may cancel their ticket and be refunded; absent: not refundable on request",
            "format" : "date-time"
          },
          "tickets" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PublicTicket"
            }
          }
        }
      },
      "PublicTicket" : {
        "required" : [ "id", "label", "price", "remaining" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "label" : {
            "type" : "string"
          },
          "price" : {
            "type" : "number",
            "description" : "Price in euros"
          },
          "remaining" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}